+++ src/primitives/primitive-borderless-button.component.jsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 const PrimitiveBorderlessButton = styled.button`
   background: transparent;
   border: none;
-  padding: 0; 
+  padding: 0;
   margin: 0;
   &&& {
     outline: none;