+++ test/component/container/container.component.spec.jsx
@@ -1,11 +1,13 @@
+/* eslint-disable prefer-arrow-callback */
+/* eslint-disable func-names */
 /* eslint-disable function-paren-newline */
 /* import React from 'react';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
 import { Container } from '../../../src/index'; */
 
-describe('Container component', () => {
-  it('should render correctly'); /* , () => {
+describe('Container component'); /* , function () {
+  it('should render correctly', () => {
     const wrapper = mount(
       <Container containerId="Example1" containerCls="base-container" >
         <div id="header-of-container">Header</div>
@@ -15,5 +17,5 @@ describe('Container component', () => {
 
     expect(wrapper.find('#header-of-container').text()).to.eql('Header');
     expect(wrapper.find('#content-of-container').text()).to.eql('Content');
-  }); */
-});
+  });
+}); */