+++ test/register.js
@@ -1,6 +1,5 @@
-require('babel-register')({
-  ignore: /node_modules/,
-});
+require('@babel/register')();
+require('raf/polyfill');
 require('ignore-styles');
 require('global-jsdom')(undefined, {
   url: 'http://localhost',