+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 ## <next>
 
 ## 3.5.2
-* Allow `styled-components` 5 in peerDependencies
-* Upgrade some packages
+* Add more styled-components 5 compability
+* Fix `Content.ReadonlyRow` id rendering
+* Upgrade to react 16.8.6
+* Upgrade packages
 
 ## 3.5.1
 * added missing semicolons to primitive button styles (fixes an issue with styled-components 5.0.0)